Enforcing consistent naming conventions across all refactored Vlasov app tests.
