+++ regression/rt_vlasov_lbo_cross_1x1v_p2.c
@@ -19,34 +19,37 @@
 
 #include <rt_arg_parse.h>
 
-struct lbo_relax_ctx
+struct lbo_cross_ctx
 {
   // Mathematical constants (dimensionless).
   double pi;
 
   // Physical constants (using normalized code units).
-  double mass; // Top hat/bump mass.
-  double charge; // Top hat/bump charge. 
+  double mass_neut1; // First neutral mass.
+  double mass_neut2; // Second neutral mass.
+  double charge_neut1; // First neutral charge.
+  double charge_neut2; // Second neutral charge.
 
-  double n0; // Reference number density.
-  double ux0; // Reference velocity (x-direction).
-  double uy0; // Reference velocity (y-direction).
-  double vt; // Top hat Maxwellian thermal velocity.
-  double nu; // Collision frequency.
+  double p_neut1; // First neutral pressure.
+  double p_neut2; // Second neutral pressure.
+  double n0_neut1; // First neutral reference number density.
+  double n0_neut2; // Second netural reference number density.
 
-  double ab; // Bump Maxwellian amplitude.
-  double sb; // Bump Maxwellian softening factor, to avoid divergence.
-  double ubx; // Bump location (x-direction, in velocity space).
-  double uby; // Bump location (y-direction, in velocity space).
-  double vtb; // Bump Maxwellian thermal velocity.
+  double ux0_neut1; // First neutral reference velocity (x-direction).
+  double ux0_neut2; // Second neutral reference velocity (x-direction).
+  double nu_neut1; // First neutral collision frequency.
+  double nu_neut2; // Second neutral collision frequency.
+
+  // Derived physical quantities (using normalized code units).
+  double vt_neut1; // First neutral thermal velocity.
+  double vt_neut2; // Second neutral thermal velocity.
 
   // Simulation parameters.
   int Nx; // Cell count (configuration space: x-direction).
   int Nvx; // Cell count (velocity space: vx-direction).
-  int Nvy; // Cell count (velocity space: vy-direction).
   double Lx; // Domain size (configuration space: x-direction).
-  double vx_max; // Domain boundary (velocity space: vx-direction).
-  double vy_max; // Domain boundary (velocity space: vy-direction).
+  double vx_max_neut1; // First neutral domain boundary (velocity space: vx-direction).
+  double vx_max_neut2; // Second neutral domain boundary (velocity space: vx-direction).
   int poly_order; // Polynomial order.
   double cfl_frac; // CFL coefficient.
 
@@ -56,63 +59,67 @@ struct lbo_relax_ctx
   int num_failures_max; // Maximum allowable number of consecutive small time-steps.
 };
 
-struct lbo_relax_ctx
+struct lbo_cross_ctx
 create_ctx(void)
 {
   // Mathematical constants (dimensionless).
   double pi = M_PI;
 
   // Physical constants (using normalized code units).
-  double mass = 1.0; // Top hat/bump mass.
-  double charge = 0.0; // Top hat/bump charge.
+  double mass_neut1 = 1.0; // First neutral mass.
+  double mass_neut2 = 0.05; // Second neutral mass.
+  double charge_neut1 = 0.0; // First neutral charge.
+  double charge_neut2 = 0.0; // Second neutral charge.
 
-  double n0 = 1.0; // Reference number density.
-  double ux0 = 0.0; // Reference velocity (x-direction).
-  double uy0 = 0.0; // Reference velocity (y-direction).
-  double vt = 1.0 / 3.0; // Top hat Maxwellian thermal velocity.
-  double nu = 0.01; // Collision frequency.
+  double p_neut1 = 1.0; // First neutral pressure.
+  double p_neut2 = 0.5; // Second neutral pressure.
+  double n0_neut1 = 1.0; // First neutral reference number density.
+  double n0_neut2 = 1.0; // Second netural reference number density.
 
-  double ab = sqrt(0.1); // Bump Maxwellian amplitude.
-  double sb = 0.12; // Bump Maxwellian softening factor, to avoid divergence.
-  double ubx = 4.0 * sqrt(0.25 / 3.0); // Bump location (x-direction, in velocity space).
-  double uby = 0.0; // Bump location (y-direction, in velocity space).
-  double vtb = 1.0; // Bump Maxwellian thermal velocity.
+  double ux0_neut1 = 0.1; // First neutral reference velocity (x-direction).
+  double ux0_neut2 = 2.5; // Second neutral reference velocity (x-direction).
+  double nu_neut1 = 1.0 / 0.01; // First neutral collision frequency.
+  double nu_neut2 = sqrt(0.5 / 0.05) / 0.01; // Second neutral collision frequency.
+
+  // Derived physical quantities (using normalized code units).
+  double vt_neut1 = sqrt(p_neut1 / (n0_neut1 * mass_neut1)); // First neutral thermal velocity.
+  double vt_neut2 = sqrt(p_neut2 / (n0_neut2 * mass_neut2)); // Second neutral thermal velocity.
 
   // Simulation parameters.
-  int Nx = 2; // Cell count (configuration space: x-direction).
-  int Nvx = 16; // Cell count (velocity space: vx-direction).
-  int Nvy = 16; // Cell count (velocity space: vy-direction).
+  int Nx = 16; // Cell count (configuration space: x-direction).
+  int Nvx = 32; // Cell count (velocity space: vx-direction).
   double Lx = 1.0; // Domain size (configuration space: x-direction).
-  double vx_max = 8.0 * vt; // Domain boundary (velocity space: vx-direction).
-  double vy_max = 8.0 * vt; // Domain boundary (velocity space: vy-direction).
-  int poly_order = 1; // Polynomial order.
+  double vx_max_neut1 = 6.0 * vt_neut1; // First neutral domain boundary (velocity space: vx-direction).
+  double vx_max_neut2 = 6.0 * vt_neut2; // Second neutral domain boundary (velocity space: vx-direction).
+  int poly_order = 2; // Polynomial order.
   double cfl_frac = 1.0; // CFL coefficient.
 
-  double t_end = 50.0; // Final simulation time.
+  double t_end = 0.0025; // Final simulation time.
   int num_frames = 1; // Number of output frames.
   double dt_failure_tol = 1.0e-4; // Minimum allowable fraction of initial time-step.
   int num_failures_max = 20; // Maximum allowable number of consecutive small time-steps.
   
-  struct lbo_relax_ctx ctx = {
+  struct lbo_cross_ctx ctx = {
     .pi = pi,
-    .mass = mass,
-    .charge = charge,
-    .n0 = n0,
-    .ux0 = ux0,
-    .uy0 = uy0,
-    .vt = vt,
-    .nu = nu,
-    .ab = ab,
-    .sb = sb,
-    .ubx = ubx,
-    .uby = uby,
-    .vtb = vtb,
+    .mass_neut1 = mass_neut1,
+    .mass_neut2 = mass_neut2,
+    .charge_neut1 = charge_neut1,
+    .charge_neut2 = charge_neut2,
+    .p_neut1 = p_neut1,
+    .p_neut2 = p_neut2,
+    .n0_neut1 = n0_neut1,
+    .n0_neut2 = n0_neut2,
+    .ux0_neut1 = ux0_neut1,
+    .ux0_neut2 = ux0_neut2,
+    .nu_neut1 = nu_neut1,
+    .nu_neut2 = nu_neut2,
+    .vt_neut1 = vt_neut1,
+    .vt_neut2 = vt_neut2,
     .Nx = Nx,
     .Nvx = Nvx,
-    .Nvy = Nvy,
     .Lx = Lx,
-    .vx_max = vx_max,
-    .vy_max = vy_max,
+    .vx_max_neut1 = vx_max_neut1,
+    .vx_max_neut2 = vx_max_neut2,
     .poly_order = poly_order,
     .cfl_frac = cfl_frac,
     .t_end = t_end,
@@ -125,61 +132,61 @@ create_ctx(void)
 }
 
 void
-evalTopHatInit(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
+evalNeut1Init(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
 {
-  struct lbo_relax_ctx *app = ctx;
-  double vx = xn[1], vy = xn[2];
+  struct lbo_cross_ctx *app = ctx;
+  double v = xn[1];
   
-  double n0 = app->n0;
-
-  double dist = 0.0;
-
-  if(fabs(vx) < 1.0 && fabs(vy) < 1.0) {
-    dist = 0.5 * n0;
-  }
-  else {
-    dist = 0.0;
-  }
-
-  // Set distribution function.
-  fout[0] = dist;
-}
-
-void
-evalBumpInit(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
-{
-  struct lbo_relax_ctx *app = ctx;
-  double vx = xn[1], vy = xn[2];
-
   double pi = app->pi;
 
-  double n0 = app->n0;
-  double ux0 = app->ux0;
-  double uy0 = app->uy0;
-  double vt = app->vt;
+  double n0_neut1 = app->n0_neut1;
+  double ux0_neut1 = app->ux0_neut1;
+  double vt_neut1 = app->vt_neut1;
 
-  double ab = app->ab;
-  double sb = app->sb;  
-  double ubx = app->ubx;
-  double uby = app->uby;
-  double vtb = app->vtb;
-
-  double v_sq = ((vx - ux0) * (vx - ux0)) + ((vy - uy0) * (vy - uy0));
-  double vb_sq = ((vx - ubx) * (vx - ubx)) + ((vy - uby) * (vy - uby));
+  double v_sq = (v - ux0_neut1) * (v - ux0_neut1);
 
   // Set distribution function.
-  fout[0] = (n0 / sqrt(2.0 * pi * vt * vt)) * exp(-v_sq / (2.0 * vt * vt)) + (n0 / sqrt(2.0 * pi * vtb * vtb)) * exp(-vb_sq / (2.0 * vtb * vtb)) * (ab * ab) / (vb_sq + (sb * sb));
+  fout[0] = (n0_neut1 / sqrt(2.0 * pi * vt_neut1 * vt_neut1)) * exp(-v_sq / (2.0 * vt_neut1 * vt_neut1));
 }
 
 void
-evalNu(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
+evalNeut2Init(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
 {
-  struct lbo_relax_ctx *app = ctx;
+  struct lbo_cross_ctx *app = ctx;
+  double v = xn[1];
+  
+  double pi = app->pi;
 
-  double nu = app->nu;
+  double n0_neut2 = app->n0_neut2;
+  double ux0_neut2 = app->ux0_neut2;
+  double vt_neut2 = app->vt_neut2;
+
+  double v_sq = (v - ux0_neut2) * (v - ux0_neut2);
+
+  // Set distribution function.
+  fout[0] = (n0_neut2 / sqrt(2.0 * pi * vt_neut2 * vt_neut2)) * exp(-v_sq / (2.0 * vt_neut2 * vt_neut2));
+}
+
+void
+evalNeut1Nu(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
+{
+  struct lbo_cross_ctx *app = ctx;
+
+  double nu_neut1 = app->nu_neut1;
 
   // Set collision frequency.
-  fout[0] = nu;
+  fout[0] = nu_neut1;
+}
+
+void
+evalNeut2Nu(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
+{
+  struct lbo_cross_ctx *app = ctx;
+
+  double nu_neut2 = app->nu_neut2;
+
+  // Set collision frequency.
+  fout[0] = nu_neut2;
 }
 
 void
@@ -214,11 +221,10 @@ main(int argc, char **argv)
     gkyl_mem_debug_set(true);
   }
 
-  struct lbo_relax_ctx ctx = create_ctx(); // Context for initialization functions.
+  struct lbo_cross_ctx ctx = create_ctx(); // Context for initialization functions.
 
   int NX = APP_ARGS_CHOOSE(app_args.xcells[0], ctx.Nx);
   int NVX = APP_ARGS_CHOOSE(app_args.vcells[0], ctx.Nvx);
-  int NVY = APP_ARGS_CHOOSE(app_args.vcells[1], ctx.Nvy);
 
   int nrank = 1; // Number of processors in simulation.
 #ifdef GKYL_HAVE_MPI
@@ -306,46 +312,50 @@ main(int argc, char **argv)
     goto mpifinalize;
   }
 
-  // Top hat species.
-  struct gkyl_vlasov_species square = {
-    .name = "square",
-    .charge = ctx.charge, .mass = ctx.mass,
-    .lower = { -ctx.vx_max, -ctx.vy_max },
-    .upper = { ctx.vx_max, ctx.vy_max }, 
-    .cells = { NVX, NVY },
+  // First neutral species.
+  struct gkyl_vlasov_species neut1 = {
+    .name = "neut1",
+    .charge = ctx.charge_neut1, .mass = ctx.mass_neut1,
+    .lower = { -ctx.vx_max_neut1 },
+    .upper = { ctx.vx_max_neut1 }, 
+    .cells = { NVX },
 
     .projection = {
       .proj_id = GKYL_PROJ_FUNC,
-      .func = evalTopHatInit,
+      .func = evalNeut1Init,
       .ctx_func = &ctx,
     },
     .collisions =  {
       .collision_id = GKYL_LBO_COLLISIONS,
-      .self_nu = evalNu,
+      .self_nu = evalNeut1Nu,
       .ctx = &ctx,
+      .num_cross_collisions = 1,
+      .collide_with = { "neut2" },
     },
     
     .num_diag_moments = 3,
     .diag_moments = { "M0", "M1i", "M2" },
   };
 
-  // Bump species.
-  struct gkyl_vlasov_species bump = {
-    .name = "bump",
-    .charge = ctx.charge, .mass = ctx.mass,
-    .lower = { -ctx.vx_max, -ctx.vy_max },
-    .upper = { ctx.vx_max, ctx.vy_max }, 
-    .cells = { NVX, NVY },
+  // Second neutral species.
+  struct gkyl_vlasov_species neut2 = {
+    .name = "neut2",
+    .charge = ctx.charge_neut2, .mass = ctx.mass_neut2,
+    .lower = { -ctx.vx_max_neut2 },
+    .upper = { ctx.vx_max_neut2 }, 
+    .cells = { NVX },
 
     .projection = {
       .proj_id = GKYL_PROJ_FUNC,
-      .func = evalBumpInit,
+      .func = evalNeut2Init,
       .ctx_func = &ctx,
     },
     .collisions =  {
       .collision_id = GKYL_LBO_COLLISIONS,
-      .self_nu = evalNu,
+      .self_nu = evalNeut2Nu,
       .ctx = &ctx,
+      .num_cross_collisions = 1,
+      .collide_with = { "neut1" },
     },
     
     .num_diag_moments = 3,
@@ -354,9 +364,9 @@ main(int argc, char **argv)
 
     // Vlasov-Maxwell app.
   struct gkyl_vm app_inp = {
-    .name = "lbo_vlasov_relax_1x2v_p1",
+    .name = "vlasov_lbo_cross_1x1v_p2",
 
-    .cdim = 1, .vdim = 2,
+    .cdim = 1, .vdim = 1,
     .lower = { 0.0 },
     .upper = { ctx.Lx },
     .cells = { NX },
@@ -369,7 +379,7 @@ main(int argc, char **argv)
     .periodic_dirs = { 0 },
 
     .num_species = 2,
-    .species = { square, bump },
+    .species = { neut1, neut2 },
 
     .skip_field = true,
