+++ regression/rt_vlasov_neut_bgk_sodshock_1x1v_p2.c
@@ -6,6 +6,7 @@
 #include <gkyl_alloc.h>
 #include <gkyl_vlasov.h>
 #include <gkyl_util.h>
+#include <gkyl_wv_euler.h>
 
 #include <gkyl_null_comm.h>
 
@@ -19,37 +20,27 @@
 
 #include <rt_arg_parse.h>
 
-struct lbo_cross_ctx
+struct sodshock_ctx
 {
-  // Mathematical constants (dimensionless).
-  double pi;
-
   // Physical constants (using normalized code units).
-  double mass_neut1; // First neutral mass.
-  double mass_neut2; // Second neutral mass.
-  double charge_neut1; // First neutral charge.
-  double charge_neut2; // Second neutral charge.
+  double mass; // Neutral mass.
+  double charge; // Neutral charge.
 
-  double p_neut1; // First neutral pressure.
-  double p_neut2; // Second neutral pressure.
-  double n0_neut1; // First neutral reference number density.
-  double n0_neut2; // Second netural reference number density.
+  double nl; // Left number density.
+  double Tl; // Left temperature.
 
-  double ux0_neut1; // First neutral reference velocity (x-direction).
-  double ux0_neut2; // Second neutral reference velocity (x-direction).
-  double nu_neut1; // First neutral collision frequency.
-  double nu_neut2; // Second neutral collision frequency.
+  double nr; // Right number density.
+  double Tr; // Right temperature.
 
-  // Derived physical quantities (using normalized code units).
-  double vt_neut1; // First neutral thermal velocity.
-  double vt_neut2; // Second neutral thermal velocity.
+  double vt; // Thermal velocity.
+  double Vx_drift; // Drift velocity (x-direction).
+  double nu; // Collision frequency.
 
   // Simulation parameters.
   int Nx; // Cell count (configuration space: x-direction).
   int Nvx; // Cell count (velocity space: vx-direction).
   double Lx; // Domain size (configuration space: x-direction).
-  double vx_max_neut1; // First neutral domain boundary (velocity space: vx-direction).
-  double vx_max_neut2; // Second neutral domain boundary (velocity space: vx-direction).
+  double vx_max; // Domain boundary (velocity space: vx-direction).
   int poly_order; // Polynomial order.
   double cfl_frac; // CFL coefficient.
 
@@ -59,67 +50,50 @@ struct lbo_cross_ctx
   int num_failures_max; // Maximum allowable number of consecutive small time-steps.
 };
 
-struct lbo_cross_ctx
+struct sodshock_ctx
 create_ctx(void)
 {
-  // Mathematical constants (dimensionless).
-  double pi = M_PI;
-
   // Physical constants (using normalized code units).
-  double mass_neut1 = 1.0; // First neutral mass.
-  double mass_neut2 = 0.05; // Second neutral mass.
-  double charge_neut1 = 0.0; // First neutral charge.
-  double charge_neut2 = 0.0; // Second neutral charge.
+  double mass = 1.0; // Neutral mass.
+  double charge = 0.0; // Neutral charge.
 
-  double p_neut1 = 1.0; // First neutral pressure.
-  double p_neut2 = 0.5; // Second neutral pressure.
-  double n0_neut1 = 1.0; // First neutral reference number density.
-  double n0_neut2 = 1.0; // Second netural reference number density.
+  double nl = 1.0; // Left number density.
+  double Tl = 1.0; // Left temperature.
 
-  double ux0_neut1 = 0.1; // First neutral reference velocity (x-direction).
-  double ux0_neut2 = 2.5; // Second neutral reference velocity (x-direction).
-  double nu_neut1 = 1.0 / 0.01; // First neutral collision frequency.
-  double nu_neut2 = sqrt(0.5 / 0.05) / 0.01; // Second neutral collision frequency.
+  double nr = 0.125; // Right number density.
+  double Tr = sqrt(0.1 / 0.125); // Right temperature.
 
-  // Derived physical quantities (using normalized code units).
-  double vt_neut1 = sqrt(p_neut1 / (n0_neut1 * mass_neut1)); // First neutral thermal velocity.
-  double vt_neut2 = sqrt(p_neut2 / (n0_neut2 * mass_neut2)); // Second neutral thermal velocity.
+  double vt = 1.0; // Thermal velocity.
+  double Vx_drift = 0.0; // Drift velocity (x-direction).
+  double nu = 100.0; // Collision frequency.
 
   // Simulation parameters.
-  int Nx = 16; // Cell count (configuration space: x-direction).
+  int Nx = 128; // Cell count (configuration space: x-direction).
   int Nvx = 32; // Cell count (velocity space: vx-direction).
   double Lx = 1.0; // Domain size (configuration space: x-direction).
-  double vx_max_neut1 = 6.0 * vt_neut1; // First neutral domain boundary (velocity space: vx-direction).
-  double vx_max_neut2 = 6.0 * vt_neut2; // Second neutral domain boundary (velocity space: vx-direction).
+  double vx_max = 8.0 * vt; // Domain boundary (velocity space: vx-direction).
   int poly_order = 2; // Polynomial order.
   double cfl_frac = 1.0; // CFL coefficient.
 
-  double t_end = 0.0025; // Final simulation time.
+  double t_end = 0.1; // Final simulation time.
   int num_frames = 1; // Number of output frames.
   double dt_failure_tol = 1.0e-4; // Minimum allowable fraction of initial time-step.
   int num_failures_max = 20; // Maximum allowable number of consecutive small time-steps.
-  
-  struct lbo_cross_ctx ctx = {
-    .pi = pi,
-    .mass_neut1 = mass_neut1,
-    .mass_neut2 = mass_neut2,
-    .charge_neut1 = charge_neut1,
-    .charge_neut2 = charge_neut2,
-    .p_neut1 = p_neut1,
-    .p_neut2 = p_neut2,
-    .n0_neut1 = n0_neut1,
-    .n0_neut2 = n0_neut2,
-    .ux0_neut1 = ux0_neut1,
-    .ux0_neut2 = ux0_neut2,
-    .nu_neut1 = nu_neut1,
-    .nu_neut2 = nu_neut2,
-    .vt_neut1 = vt_neut1,
-    .vt_neut2 = vt_neut2,
+
+  struct sodshock_ctx ctx = {
+    .mass = mass,
+    .charge = charge,
+    .nl = nl,
+    .Tl = Tl,
+    .nr = nr,
+    .Tr = Tr,
+    .vt = vt,
+    .Vx_drift = Vx_drift,
+    .nu = nu,
     .Nx = Nx,
     .Nvx = Nvx,
     .Lx = Lx,
-    .vx_max_neut1 = vx_max_neut1,
-    .vx_max_neut2 = vx_max_neut2,
+    .vx_max = vx_max,
     .poly_order = poly_order,
     .cfl_frac = cfl_frac,
     .t_end = t_end,
@@ -132,61 +106,69 @@ create_ctx(void)
 }
 
 void
-evalNeut1Init(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
+evalDensityInit(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
 {
-  struct lbo_cross_ctx *app = ctx;
-  double v = xn[1];
-  
-  double pi = app->pi;
+  struct sodshock_ctx *app = ctx;
+  double x = xn[0];
 
-  double n0_neut1 = app->n0_neut1;
-  double ux0_neut1 = app->ux0_neut1;
-  double vt_neut1 = app->vt_neut1;
+  double nl = app->nl;
+  double nr = app->nr;
 
-  double v_sq = (v - ux0_neut1) * (v - ux0_neut1);
+  double n = 0.0;
+
+  if (x < 0.5) {
+    n = nl;
+  }
+  else {
+    n = nr;
+  }
 
   // Set distribution function.
-  fout[0] = (n0_neut1 / sqrt(2.0 * pi * vt_neut1 * vt_neut1)) * exp(-v_sq / (2.0 * vt_neut1 * vt_neut1));
+  fout[0] = n;
 }
 
 void
-evalNeut2Init(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
+evalTempInit(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
 {
-  struct lbo_cross_ctx *app = ctx;
-  double v = xn[1];
-  
-  double pi = app->pi;
+  struct sodshock_ctx *app = ctx;
+  double x = xn[0];
 
-  double n0_neut2 = app->n0_neut2;
-  double ux0_neut2 = app->ux0_neut2;
-  double vt_neut2 = app->vt_neut2;
+  double Tl = app->Tl;
+  double Tr = app->Tr;
 
-  double v_sq = (v - ux0_neut2) * (v - ux0_neut2);
+  double T = 0.0;
 
-  // Set distribution function.
-  fout[0] = (n0_neut2 / sqrt(2.0 * pi * vt_neut2 * vt_neut2)) * exp(-v_sq / (2.0 * vt_neut2 * vt_neut2));
+  if (x < 0.5) {
+    T = Tl;
+  }
+  else {
+    T = Tr;
+  }
+
+  // Set temperature.
+  fout[0] = T;
 }
 
 void
-evalNeut1Nu(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
+evalVDriftInit(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
 {
-  struct lbo_cross_ctx *app = ctx;
+  struct sodshock_ctx *app = ctx;
 
-  double nu_neut1 = app->nu_neut1;
+  double Vx_drift = app->Vx_drift;
+
+  // Set drift velocity.
+  fout[0] = Vx_drift;
+}
+
+void
+evalNu(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
+{
+  struct sodshock_ctx *app = ctx;
+
+  double nu = app->nu;
 
   // Set collision frequency.
-  fout[0] = nu_neut1;
-}
-
-void
-evalNeut2Nu(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void* ctx)
-{
-  struct lbo_cross_ctx *app = ctx;
-
-  double nu_neut2 = app->nu_neut2;
-
-  // Set collision frequency.
-  fout[0] = nu_neut2;
+  fout[0] = nu;
 }
 
 void
@@ -221,7 +203,7 @@ main(int argc, char **argv)
     gkyl_mem_debug_set(true);
   }
 
-  struct lbo_cross_ctx ctx = create_ctx(); // Context for initialization functions.
+  struct sodshock_ctx ctx = create_ctx(); // Context for initialization functions.
 
   int NX = APP_ARGS_CHOOSE(app_args.xcells[0], ctx.Nx);
   int NVX = APP_ARGS_CHOOSE(app_args.vcells[0], ctx.Nvx);
@@ -312,84 +294,64 @@ main(int argc, char **argv)
     goto mpifinalize;
   }
 
-  // First neutral species.
-  struct gkyl_vlasov_species neut1 = {
-    .name = "neut1",
-    .charge = ctx.charge_neut1, .mass = ctx.mass_neut1,
-    .lower = { -ctx.vx_max_neut1 },
-    .upper = { ctx.vx_max_neut1 }, 
+  // Neutral species.
+  struct gkyl_vlasov_species neut = {
+    .name = "neut",
+    .model_id = GKYL_MODEL_DEFAULT,
+    .charge = ctx.charge, .mass = ctx.mass,
+    .lower = { -ctx.vx_max },
+    .upper = { ctx.vx_max }, 
     .cells = { NVX },
 
     .projection = {
-      .proj_id = GKYL_PROJ_FUNC,
-      .func = evalNeut1Init,
-      .ctx_func = &ctx,
+      .proj_id = GKYL_PROJ_VLASOV_LTE,
+      .density = evalDensityInit,
+      .ctx_density = &ctx,
+      .temp = evalTempInit,
+      .ctx_temp = &ctx,
+      .V_drift = evalVDriftInit,
+      .ctx_V_drift = &ctx,
+      .correct_all_moms = true,
     },
     .collisions =  {
-      .collision_id = GKYL_LBO_COLLISIONS,
-      .self_nu = evalNeut1Nu,
+      .collision_id = GKYL_BGK_COLLISIONS,
+      .self_nu = evalNu,
       .ctx = &ctx,
-      .num_cross_collisions = 1,
-      .collide_with = { "neut2" },
+      .correct_all_moms = true,
     },
     
     .num_diag_moments = 3,
-    .diag_moments = { "M0", "M1i", "M2" },
+    .diag_moments = { "M0", "M1i", "LTEMoments" },
   };
 
-  // Second neutral species.
-  struct gkyl_vlasov_species neut2 = {
-    .name = "neut2",
-    .charge = ctx.charge_neut2, .mass = ctx.mass_neut2,
-    .lower = { -ctx.vx_max_neut2 },
-    .upper = { ctx.vx_max_neut2 }, 
-    .cells = { NVX },
-
-    .projection = {
-      .proj_id = GKYL_PROJ_FUNC,
-      .func = evalNeut2Init,
-      .ctx_func = &ctx,
-    },
-    .collisions =  {
-      .collision_id = GKYL_LBO_COLLISIONS,
-      .self_nu = evalNeut2Nu,
-      .ctx = &ctx,
-      .num_cross_collisions = 1,
-      .collide_with = { "neut1" },
-    },
-    
-    .num_diag_moments = 3,
-    .diag_moments = { "M0", "M1i", "M2" },
-  };
-
-    // Vlasov-Maxwell app.
+  // Vlasov-Maxwell app.
   struct gkyl_vm app_inp = {
-    .name = "lbo_vlasov_cross_1x1v_p2",
+   .name = "vlasov_neut_bgk_sodshock_1x1v_p2",
 
-    .cdim = 1, .vdim = 1,
-    .lower = { 0.0 },
-    .upper = { ctx.Lx },
-    .cells = { NX },
+   .cdim = 1, .vdim = 1, 
+   .lower = { 0.0 },
+   .upper = { ctx.Lx },
+   .cells = { NX },
 
-    .poly_order = ctx.poly_order,
-    .basis_type = app_args.basis_type,
-    .cfl_frac = ctx.cfl_frac,
+   .poly_order = ctx.poly_order,
+   .basis_type = app_args.basis_type,
+   .cfl_frac = ctx.cfl_frac,
 
-    .num_periodic_dir = 1,
-    .periodic_dirs = { 0 },
+   .num_periodic_dir = 0,
+   .periodic_dirs = { },
 
-    .num_species = 2,
-    .species = { neut1, neut2 },
+   .num_species = 1,
+   .species = { neut },
 
-    .skip_field = true,
+   .skip_field = true,
 
-    .use_gpu = app_args.use_gpu,
+   .use_gpu = app_args.use_gpu,
 
-    .has_low_inp = true,
-    .low_inp = {
-      .local_range = decomp->ranges[my_rank],
-      .comm = comm
-    }
+   .has_low_inp = true,
+   .low_inp = {
+     .local_range = decomp->ranges[my_rank],
+     .comm = comm
+   }
   };
 
   // Create app object.
